feat(pixiu): add audit_cleaner default
